% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coloring.R
\name{add_cell_coloring}
\alias{add_cell_coloring}
\title{Add coloring}
\usage{
add_cell_coloring(cell_positions, color_cells = c("auto", "none", "grouping",
  "feature", "milestone", "pseudotime"), traj, grouping_assignment = NULL,
  groups = NULL, feature_oi = NULL, expression_source = "expression",
  pseudotime = NULL, color_milestones = c("auto", "given"),
  milestones = NULL, milestone_percentages = NULL)
}
\arguments{
\item{cell_positions}{The positions of the cells}

\item{color_cells}{How to color the cells}

\item{traj}{The trajectory}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{pseudotime}{The pseudotime}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}
}
\description{
Add coloring
}
