% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_graph.R
\name{add_cell_graph}
\alias{add_cell_graph}
\title{Add a cell graph trajectory to a data wrapper}
\usage{
add_cell_graph(data_wrapper, cell_graph, to_keep,
  milestone_prefix = "milestone_", ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{cell_graph}{The edges between cells. Format: Data frame(from = character, to = character, length = numeric)}

\item{to_keep}{A named vector containing booleans containing
whether or not a sample is part of the trajectory that is to be kept.}

\item{milestone_prefix}{A prefix to add to the id of the cell ids when they are used as milestones, in order to avoid any naming conflicts,}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
