% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_waypoints.R
\name{add_cell_waypoints}
\alias{add_cell_waypoints}
\title{Add cell waypoints to a wrapped object with trajectory}
\usage{
add_cell_waypoints(object, num_cells_selected = 100)
}
\arguments{
\item{object}{Wrapper with trajectory}

\item{num_cells_selected}{About the number of cells selected as waypoints}
}
\description{
Add cell waypoints to a wrapped object with trajectory
}
