% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cluster_graph_to_wrapper.R
\name{add_cluster_graph_to_wrapper}
\alias{add_cluster_graph_to_wrapper}
\title{Add a cluster projection trajectory to a data wrapper}
\usage{
add_cluster_graph_to_wrapper(data_wrapper, milestone_network, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon. Needs to have a cell grouping created by \code{\link{add_grouping_to_wrapper}}.}

\item{milestone_network}{A network of milestones.}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
