% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coloring.R
\name{add_density_coloring}
\alias{add_density_coloring}
\title{Color cells using a background density}
\usage{
add_density_coloring(cell_positions, color_density = c("none", "grouping",
  "feature"), traj, grouping_assignment = NULL, groups = NULL,
  feature_oi = NULL, expression_source = "expression", padding = 0.1,
  nbins = 1000, bw = 0.2, density_cutoff = 0.3,
  density_cutoff_label = density_cutoff/10)
}
\arguments{
\item{cell_positions}{The positions of the cells in 2D}

\item{color_density}{How to color density, can be "none", "grouping", or "feature"}

\item{traj}{The trajectory}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{padding}{The padding in the edges to the plot, relative to the size of the plot}

\item{nbins}{Number of bins for calculating the density}

\item{bw}{Bandwidth, relative to the size of the plot}

\item{density_cutoff}{Cutoff for density, the lower the larger the areas}

\item{density_cutoff_label}{Cutoff for density for labelling, the lower the further way from cells}
}
\description{
Color cells using a background density
}
