% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred.R
\name{add_dimred}
\alias{add_dimred}
\alias{is_wrapper_with_dimred}
\alias{get_dimred}
\title{Add a dimensionality reduction to a data wrapper}
\usage{
add_dimred(data_wrapper, dimred, dimred_milestones = NULL,
  dimred_trajectory_segments = NULL, expression_source = "expression", ...)

is_wrapper_with_dimred(data_wrapper)

get_dimred(data_wrapper, dimred = NULL, expression_source = "expression")
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{dimred_milestones}{An optional dimensionality reduction of the milestones.}

\item{dimred_trajectory_segments}{An optional dimensionality reduction of the trajectory segments.}

\item{expression_source}{The source of expression, can be "counts", "expression", an expression matrix, or another data wrapper which contains expression}

\item{...}{extra information to be stored in the wrapper}
}
\description{
TODO: add possibility to also dimred the milestones and segments
}
