% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred_projection.R
\name{add_dimred_projection}
\alias{add_dimred_projection}
\title{Add a dimred projection trajectory to a data wrapper}
\usage{
add_dimred_projection(data_wrapper, milestone_ids, milestone_network, dimred,
  dimred_milestones, grouping = NULL, num_segments_per_edge = 100, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{dimred_milestones}{An optional dimensionality reduction of the milestones.}

\item{grouping}{A grouping of the cells.}

\item{num_segments_per_edge}{The number of pieces each milestone transition will be split info for the calculation of the percentages.}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
