% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_expression.R
\name{add_expression}
\alias{add_expression}
\alias{is_wrapper_with_expression}
\alias{get_expression}
\title{Add count and normalised expression values to a data wrapper}
\usage{
add_expression(data_wrapper, counts, expression, feature_info = NULL, ...)

is_wrapper_with_expression(data_wrapper)

get_expression(data_wrapper, expression_source = "expression")
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon}

\item{counts}{The counts with genes in columns and cells in rows}

\item{expression}{The normalised expression values with genes in columns and cells in rows}

\item{feature_info}{Optional meta-information of the features, a data.frame with at least feature_id as column}

\item{...}{extra information to be stored in the wrapper}

\item{expression_source}{The source of expression, can be "counts", "expression", an expression matrix, or another data wrapper which contains expression}
}
\description{
Add count and normalised expression values to a data wrapper
}
