% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_grouping.R
\name{add_grouping}
\alias{add_grouping}
\title{Add a cell grouping to a data wrapper}
\usage{
add_grouping(data_wrapper, group_ids, grouping, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{group_ids}{The ids of the groupings.}

\item{grouping}{A grouping of the cells.}

\item{...}{Extra information to be stored in the wrapper.}
}
\description{
Add a cell grouping to a data wrapper
}
