% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_linear_trajectory.R
\name{add_linear_trajectory}
\alias{add_linear_trajectory}
\title{Add a linear trajectory to a data wrapper}
\usage{
add_linear_trajectory(data_wrapper, pseudotimes, directed = FALSE,
  do_scale_minmax = TRUE, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{pseudotimes}{A named vector of pseudo times.}

\item{directed}{Whether or not the directionality of the pseudotime is predicted.}

\item{do_scale_minmax}{Whether or not to scale the pseudotimes between [0,1].
Otherwise, will assume the values are already within that range.}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
