% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coloring.R
\name{add_milestone_coloring}
\alias{add_milestone_coloring}
\title{Add milestone coloring}
\usage{
add_milestone_coloring(milestones = NULL, color_milestones = c("auto",
  "given"))
}
\arguments{
\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{color_milestones}{How to color the cells}
}
\description{
Add milestone coloring
}
