% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_prior_information.R
\name{add_prior_information}
\alias{add_prior_information}
\title{Add prior information to a data wrapper}
\usage{
add_prior_information(task, start_cells = NULL, end_cells = NULL,
  grouping_assignment = NULL, grouping_network = NULL,
  marker_feature_ids = NULL, n_branches = NULL, n_end_states = NULL,
  time = NULL)
}
\arguments{
\item{task}{A data wrapper to extend upon.}

\item{start_cells}{The start cells}

\item{end_cells}{The end cells}

\item{grouping_assignment}{The grouping of cells, a dataframe with cell_id and group_id}

\item{grouping_network}{The network between groups, a dataframe with from and to}

\item{marker_feature_ids}{The features (genes) important for the trajectory}

\item{n_branches}{Number of branches}

\item{n_end_states}{Number of end states}

\item{time}{The time for every cell}
}
\description{
Note that the given data wrapper requires a trajectory and expression values
to have been added already.
}
