% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_prior_information_to_wrapper.R
\name{add_prior_information_to_wrapper}
\alias{add_prior_information_to_wrapper}
\title{Add prior information to a data wrapper}
\usage{
add_prior_information_to_wrapper(data_wrapper)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}
}
\description{
Note that the given data wrapper requires a trajectory and expression values
to have been added already.
}
