% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_pseudotime.R
\name{calculate_pseudotime}
\alias{calculate_pseudotime}
\alias{add_pseudotime}
\title{Calculate global pseudotime as distance from root}
\usage{
calculate_pseudotime(trajectory)

add_pseudotime(trajectory)
}
\arguments{
\item{trajectory}{The trajectory}
}
\description{
Calculate global pseudotime as distance from root
}
