% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_root.R
\name{add_root}
\alias{add_root}
\title{Root the trajectory}
\usage{
add_root(trajectory, root_cell_id = trajectory$root_cell_id,
  root_milestone_id = trajectory$root_milestone_id)
}
\arguments{
\item{trajectory}{the trajectory object}

\item{root_cell_id}{The root cell id, not required if root_milestone_id is given}

\item{root_milestone_id}{The root milestone id, not required if root_cell_id is given}
}
\description{
Roots the trajectory by changing the directionality of all edges given a root cell
}
