% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_root.R
\name{add_root_using_expression}
\alias{add_root_using_expression}
\title{Add root cell to wrapper using expression of features}
\usage{
add_root_using_expression(trajectory, features_oi,
  expression_source = "expression")
}
\arguments{
\item{trajectory}{the trajectory object}

\item{features_oi}{The feature ids which will be used to root}

\item{expression_source}{Source of the expression, either a string or a matrix}
}
\description{
Add root cell to wrapper using expression of features
}
