% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_timings_to_wrapper.R
\name{add_timing_checkpoint}
\alias{add_timing_checkpoint}
\title{Helper function for storing timings information.}
\usage{
add_timing_checkpoint(timings, name)
}
\arguments{
\item{timings}{The timings list of previous checkpoints.}

\item{name}{The name of the timings checkpoint.}
}
\description{
Helper function for storing timings information.
}
