% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_timings.R
\name{add_timings}
\alias{add_timings}
\title{Add count and normalised expression values to a data wrapper}
\usage{
add_timings(data_wrapper, timings)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{timings}{A list of timings.}
}
\description{
Add count and normalised expression values to a data wrapper
}
