% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_trajectory_to_wrapper.R
\name{add_trajectory_to_wrapper}
\alias{add_trajectory_to_wrapper}
\title{Add a trajectory to a data wrapper}
\usage{
add_trajectory_to_wrapper(data_wrapper, milestone_ids, milestone_network,
  divergence_regions, milestone_percentages = NULL, progressions = NULL,
  ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{divergence_regions}{A data frame specifying the divergence
regions between milestones (e.g. a bifurcation).
Type: Data frame(divergence_id = character, milestone_id = character, is_start = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}

\item{progressions}{Specifies the progression of a cell along a transition in the milestone_network.
Type: Data frame(cell_id = character, from = character, to = character, percentage = numeric).}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
Add a trajectory to a data wrapper
}
