% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_by.R
\name{calculate_average_by_group}
\alias{calculate_average_by_group}
\title{Calculate mean values per cell group}
\usage{
calculate_average_by_group(x, cell_grouping)
}
\arguments{
\item{x}{A matrix. One row for every cell; one column for every feature. The rows must be named.}

\item{cell_grouping}{A data frame denoting the grouping of the cells.
Format: \code{data_frame(cell_id = character(), group_id = character())}.}
}
\description{
Calculate mean values per cell group
}
