% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_by.R
\name{calculate_average_by_milestone_percentages}
\alias{calculate_average_by_milestone_percentages}
\title{Calculate mean values by milestone percentages}
\usage{
calculate_average_by_milestone_percentages(x, milestone_percentages)
}
\arguments{
\item{x}{A matrix. One row for every cell; one column for every feature. The rows must be named.}

\item{milestone_percentages}{A data frame of milestone percentages.
Format: \code{data_frame(cell_id = character(), milestone_id = character(), percentage = numeric())}.}
}
\description{
Calculate mean values by milestone percentages
}
