% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_check_dependencies.R
\name{check_dependency}
\alias{check_dependency}
\title{Check which packages need to be installed for a TI method to function correctly}
\usage{
check_dependency(descr)
}
\arguments{
\item{descr}{Description of the method}
}
\description{
Check which packages need to be installed for a TI method to function correctly
}
