% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_dimred.R
\name{check_or_perform_dimred}
\alias{check_or_perform_dimred}
\title{Perform dimred trajectory if needed}
\usage{
check_or_perform_dimred(object)
}
\arguments{
\item{object}{An object generated by \code{wrap_data} or \code{dimred_trajectory}.}
}
\description{
Perform dimred trajectory if needed
}
