% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_milestone_network.R
\name{classify_milestone_network}
\alias{classify_milestone_network}
\title{Classify a milestone network}
\usage{
classify_milestone_network(milestone_network)
}
\arguments{
\item{milestone_network}{A milestone network}
}
\description{
Classify a milestone network
}
