% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_milestone_percentages_to_progressions.R
\name{convert_milestone_percentages_to_progressions}
\alias{convert_milestone_percentages_to_progressions}
\title{Convert milestone percentages to progressions}
\usage{
convert_milestone_percentages_to_progressions(cell_ids, milestone_ids,
  milestone_network, milestone_percentages)
}
\arguments{
\item{cell_ids}{Vector of all cell ids}

\item{milestone_ids}{Vector of milestone ids}

\item{milestone_network}{Milestone network}

\item{milestone_percentages}{Milestone percentages}
}
\description{
Convert milestone percentages to progressions
}
