% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_progressions_to_milestone_percentages.R
\name{convert_progressions_to_milestone_percentages}
\alias{convert_progressions_to_milestone_percentages}
\title{Convert progressions to milestone percentages}
\usage{
convert_progressions_to_milestone_percentages(cell_ids, milestone_ids,
  milestone_network, progressions)
}
\arguments{
\item{cell_ids}{Vector of all cell ids}

\item{milestone_ids}{Vector of milestone ids}

\item{milestone_network}{Milestone network}

\item{progressions}{Progressions dataframe}
}
\description{
Convert progressions to milestone percentages
}
