% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_create_description.R
\name{create_description}
\alias{create_description}
\title{Create a TI method description}
\usage{
create_description(name, short_name, package_loaded, package_required, par_set,
  properties, run_fun, plot_fun, override_runfun_params = TRUE)
}
\arguments{
\item{name}{The name of the TI method}

\item{short_name}{A short name for the method, max 8 characters}

\item{package_loaded}{The packages that need to be loaded before executing the method}

\item{package_required}{The packages that need to be installed before executing the method}

\item{par_set}{A bunch of parameters created by \code{\link[ParamHelpers]{makeParamSet}}}

\item{properties}{Several descriptive properties of the method. WIP.}

\item{run_fun}{A function to run the TI, needs to have 'counts' as its first param.}

\item{plot_fun}{A function to plot the results of a TI, needs to have 'prediction' as its first param.}

\item{override_runfun_params}{Whether or not to override the default parameters
of \code{run_fun} with those described in \code{par_set}.}
}
\description{
Create a TI method description
}
