% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_dimred_expr.R
\name{list_dimred_methods}
\alias{list_dimred_methods}
\alias{dimred}
\title{Perform simple dimensionality reduction}
\usage{
list_dimred_methods()

dimred(x, method, ndim, ...)
}
\arguments{
\item{x}{Log transformed expression data, with rows as cells and columns as features}

\item{method}{The name of the dimensionality reduction method to use}

\item{ndim}{The number of dimensions}

\item{...}{Any arguments to be passed to the dimensionality reduction method}
}
\description{
Perform simple dimensionality reduction
}
