% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_dimred.R
\name{dimred_trajectory}
\alias{dimred_trajectory}
\title{Perform dimensionality reduction on a trajectory and the respective samples in order to plot it}
\usage{
dimred_trajectory(traj)
}
\arguments{
\item{traj}{The trajectory}
}
\description{
Perform dimensionality reduction on a trajectory and the respective samples in order to plot it
}
