% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dynguidelines}
\alias{dynguidelines}
\title{Dynguidelines packages}
\description{
The dynguidelines package can be used to find the most optimal TI methods on a given task.
This can be done both directly in R or through a shiny interface.
}
