% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fibroblast_reprogramming_treutlein}
\alias{fibroblast_reprogramming_treutlein}
\title{Dataset containing trans-differentiating fibroblasts from [Treutlein et. al.(http://doi.org/10.1038/nature18323).}
\format{An object of class \code{dynwrap::with_grouping} (inherits from \code{dynwrap::with_expression}, \code{dynwrap::data_wrapper}, \code{list}) of length 9.}
\source{
[http://doi.org/10.1038/nature18323 (http://doi.org/10.1038/nature18323)]
}
\usage{
fibroblast_reprogramming_treutlein
}
\description{
Dataset containing trans-differentiating fibroblasts from [Treutlein et. al.(http://doi.org/10.1038/nature18323).
}
\keyword{datasets}
