% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_grouping.R
\name{get_cell_grouping}
\alias{get_cell_grouping}
\title{Group cells to their highest milestone}
\usage{
get_cell_grouping(milestone_percentages)
}
\arguments{
\item{milestone_percentages}{A data frame of milestone percentages.
Format: \code{data_frame(cell_id = character(), milestone_id = character(), percentage = numeric())}.}
}
\description{
Group cells to their highest milestone
}
