% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_get_descriptions.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Return all TI descriptions}
\usage{
get_descriptions(as_tibble = TRUE)
}
\arguments{
\item{as_tibble}{Whether or not to return the descriptions as a tibble}
}
\description{
Return all TI descriptions
}
