% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_pca}
\alias{dimred_pca}
\alias{dimred_simlr}
\alias{dimred_mds}
\alias{dimred_mds_sammon}
\alias{dimred_mds_isomds}
\alias{dimred_mds_smacof}
\alias{dimred_tsne}
\alias{dimred_dp}
\alias{dimred_ica}
\alias{dimred_lle}
\alias{get_dimreds}
\title{Dimensionality reduction functions}
\usage{
dimred_pca(x, ndim = 3)

dimred_simlr(x, ndim = 3, nclusters = 4)

dimred_mds(x, ndim = 3)

dimred_mds_sammon(x, ndim = 3)

dimred_mds_isomds(x, ndim = 3)

dimred_mds_smacof(x, ndim = 3)

dimred_tsne(x, ndim = 3)

dimred_dp(x, ndim = 3, neigen = NULL)

dimred_ica(x, ndim = 3)

dimred_lle(x, ndim = 3)

get_dimreds()
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}

\item{nclusters}{Number of clusters for simlr}

\item{neigen}{Number of eigenvalues for diffusionMap}
}
\description{
Dimensionality reduction functions
}
