% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_palette.R
\name{get_milestone_palette_names}
\alias{get_milestone_palette_names}
\title{Get the names of valid color palettes}
\usage{
get_milestone_palette_names()
}
\description{
Get the names of valid color palettes
}
