% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_get_ti_methods.R
\name{get_ti_methods}
\alias{get_ti_methods}
\title{Return all TI ti_methods}
\usage{
get_ti_methods(as_tibble = TRUE)
}
\arguments{
\item{as_tibble}{Whether or not to return the ti_methods as a tibble}
}
\description{
Return all TI ti_methods
}
