% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_infer_trajectory.R
\name{infer_trajectory}
\alias{infer_trajectory}
\alias{infer_trajectories}
\title{Infer a trajectory on one (or more) tasks, using one (or more) methods}
\usage{
infer_trajectory(task, method, parameters = NULL)

infer_trajectories(tasks, methods, parameters = NULL)
}
\arguments{
\item{task}{A task, a tibble of tasks}

\item{method}{A method, list of methods, or a design tibble}

\item{parameters}{Parameters, or a list of parameters}

\item{tasks}{One or more tasks (either as a list or tibble)}

\item{methods}{One or more methods (either as a list or tibble)}
}
\description{
Infer a trajectory on one (or more) tasks, using one (or more) methods
}
