% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_check_dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install dependencies for a method}
\usage{
install_dependencies(descr)
}
\arguments{
\item{descr}{Description of the method}
}
\description{
Install dependencies for a method
}
