% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_data.R
\name{is_data_wrapper}
\alias{is_data_wrapper}
\title{Test whether an object is a data_wrapper}
\usage{
is_data_wrapper(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper
}
