% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidelines.R
\name{is_guidelines}
\alias{is_guidelines}
\title{Check whether object is guidelines}
\usage{
is_guidelines(guidelines)
}
\arguments{
\item{guidelines}{The object to check}
}
\description{
Check whether object is guidelines
}
