% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_prediction_model.R
\name{is_prediction}
\alias{is_prediction}
\title{Tests whether an object is a trajectory created by a TI method.}
\usage{
is_prediction(object)
}
\arguments{
\item{object}{The object to be tested}
}
\description{
Tests whether an object is a trajectory created by a TI method.
}
