% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_create_ti_method.R
\name{is_ti_method}
\alias{is_ti_method}
\title{Tests whether an object is a TI method description}
\usage{
is_ti_method(object)
}
\arguments{
\item{object}{The object to be tested}
}
\description{
Tests whether an object is a TI method description
}
