% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_expression_to_wrapper.R
\name{is_wrapper_with_expression}
\alias{is_wrapper_with_expression}
\title{Test whether an object is a data_wrapper and has expression data}
\usage{
is_wrapper_with_expression(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and has expression data
}
