% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_grouping.R
\name{is_wrapper_with_grouping}
\alias{is_wrapper_with_grouping}
\title{Test whether an object is a data_wrapper and has grouping data}
\usage{
is_wrapper_with_grouping(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and has grouping data
}
