% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_prior_information.R
\name{is_wrapper_with_prior_information}
\alias{is_wrapper_with_prior_information}
\title{Test whether an object is a task and contains prior information}
\usage{
is_wrapper_with_prior_information(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a task and contains prior information
}
