% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_timings_to_wrapper.R
\name{is_wrapper_with_timings}
\alias{is_wrapper_with_timings}
\title{Test whether an object is a data_wrapper and has timings information}
\usage{
is_wrapper_with_timings(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and has timings information
}
