% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_trajectory_to_wrapper.R
\name{is_wrapper_with_trajectory}
\alias{is_wrapper_with_trajectory}
\title{Test whether an object is a data_wrapper and has a trajectory}
\usage{
is_wrapper_with_trajectory(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and has a trajectory
}
