% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_waypoints.R
\name{is_wrapper_with_waypoint_cells}
\alias{is_wrapper_with_waypoint_cells}
\title{Test whether an object is a data_wrapper and cell waypoints}
\usage{
is_wrapper_with_waypoint_cells(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and cell waypoints
}
