% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_waypoints.R
\name{is_wrapper_with_waypoints}
\alias{is_wrapper_with_waypoints}
\title{Test whether an traj is a data_wrapper and waypoints}
\usage{
is_wrapper_with_waypoints(traj)
}
\arguments{
\item{traj}{The traj to be tested.}
}
\description{
Test whether an traj is a data_wrapper and waypoints
}
