% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_label_milestones.R
\name{label_milestones}
\alias{label_milestones}
\alias{is_wrapper_with_milestone_labelling}
\alias{get_milestone_labelling}
\title{Label milestones based on highest expression of a set of markers}
\usage{
label_milestones(traj, labelling, expression_source = "expression",
  n_nearest_cells = 20)

is_wrapper_with_milestone_labelling(traj)

get_milestone_labelling(traj, label_milestones = NULL)
}
\arguments{
\item{traj}{The trajectory}

\item{labelling}{List containing for each new name of a milestone the genes which will be used as markers}

\item{expression_source}{The expression source}

\item{n_nearest_cells}{The number of nearest cells to use for extracting milestone expression}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels as milestone_ids will be used), a named character vector, or FALSE}
}
\description{
Label milestones based on highest expression of a set of markers
}
