% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearise_cells.R
\name{linearise_cells}
\alias{linearise_cells}
\title{Linearise a trajectory}
\usage{
linearise_cells(milestone_network, progressions, margin = 0.05,
  no_margin_between_linear = TRUE, one_edge = FALSE,
  equal_cell_width = FALSE)
}
\arguments{
\item{milestone_network}{The milestone network}

\item{progressions}{The progressions}

\item{margin}{The margin to add}

\item{no_margin_between_linear}{Whether to add a margin only when}

\item{one_edge}{If TRUE, assigns each cell to one edge only}

\item{equal_cell_width}{if TRUE, will give each cell equal width}
}
\description{
Linearise a trajectory
}
