% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{methods}
\alias{methods}
\title{The methods and their performance as generated from `dynalysis`}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 188 columns.}
\usage{
methods
}
\description{
The methods and their performance as generated from `dynalysis`
}
\keyword{datasets}
