% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_palette.R
\name{milestone_palette}
\alias{milestone_palette}
\title{Wrapper for various palettes}
\usage{
milestone_palette(name, n)
}
\arguments{
\item{name}{The name of the palette. Must be one of \code{"cubeHelix"}, \code{"Set3"}, or \code{"rainbow"}.}

\item{n}{The number of colours to be in the palette.}
}
\description{
Wrapper for various palettes
}
