% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_paga.R
\name{ti_paga}
\alias{ti_paga}
\title{Inferring trajectories with PAGA}
\usage{
ti_paga(n_neighbours = 30L, n_comps = 50L, resolution = 2.5)
}
\arguments{
\item{n_neighbours}{Number of neighbours for knn}

\item{n_comps}{Number of principal components}

\item{resolution}{Resolution of louvain clustering, which determines the granularity of the clustering. Higher values will result in more clusters.}
}
\description{
Inferring trajectories with PAGA
}
