% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_features}
\alias{plot_features}
\title{Plotting a set of features in a line plot}
\usage{
plot_features(traj, expression_source = "expression", features_oi = 2,
  margin = 0.02, cell_feature_importances = NULL,
  scale = dynutils::scale_quantile)
}
\arguments{
\item{traj}{The trajectory}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{features_oi}{The features of interest, either the number of features or a vector giving the names of the different features}

\item{margin}{The margin to add between milestones}

\item{cell_feature_importances}{The importances of every feature in every cell, as returned by [dynfeature::calculate_cell_feature_importance()]}

\item{scale}{Whether to rescale the expression, can be a function or boolean}
}
\description{
Plotting a set of features in a line plot
}
