% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\alias{plot_default}
\title{Plot a dimensionality reduced trajectory as a 2D graph}
\usage{
plot_graph(traj, color_cells = c("auto", "none", "grouping", "feature",
  "milestone", "pseudotime"), color_milestones = c("auto", "given"),
  grouping_assignment = NULL, groups = NULL, feature_oi = NULL,
  pseudotime = NULL, expression_source = "expression", milestones = NULL,
  milestone_percentages = NULL, transition_size = 3, milestone_size = 5,
  arrow_length = grid::unit(1, "cm"), plot_label = c("leaves", "all",
  "none"), plot_milestones = FALSE)
}
\arguments{
\item{traj}{The trajectory}

\item{color_cells}{How to color the cells}

\item{color_milestones}{How to color the cells}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{transition_size}{The size of the transition lines between milestones.}

\item{milestone_size}{The size of milestones.}

\item{arrow_length}{length of the arrow.}

\item{plot_label}{What to label. Must be one of \code{"leaves"}, \code{"all"}, or \code{"none"}.}

\item{plot_milestones}{Whether to plot the milestones.}
}
\description{
Plot a dimensionality reduced trajectory as a 2D graph
}
