% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the traj as a heatmap}
\usage{
plot_heatmap(traj, expression_source = "expression", features_oi = 20,
  clust = "ward.D2", margin = 0.02, color_cells = NULL,
  milestones = NULL, milestone_percentages = traj$milestone_percentages,
  grouping_assignment = NULL, groups = NULL,
  cell_feature_importances = NULL, heatmap_type = c("tiled", "dotted"),
  scale = dynutils::scale_quantile, label_milestones = TRUE)
}
\arguments{
\item{traj}{The trajectory}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{features_oi}{The features of interest, either the number of features or a vector giving the names of the different features}

\item{clust}{The method to cluster the features, or a hclust object}

\item{margin}{The margin to add}

\item{color_cells}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{cell_feature_importances}{The importances of every feature in every cell, as returned by [dynfeature::calculate_cell_feature_importance()]}

\item{heatmap_type}{The type of heatmap, either tiled or dotted}

\item{scale}{Whether to rescale the expression, can be a function or boolean}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels as milestone_ids will be used), a named character vector, or FALSE}
}
\description{
Plot the traj as a heatmap
}
