% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strip.R
\name{plot_strip}
\alias{plot_strip}
\title{Plot strip}
\usage{
plot_strip(traj1, traj2, margin = 0.05, reorder = TRUE)
}
\arguments{
\item{traj1}{The first traj}

\item{traj2}{The second traj}

\item{margin}{The margin to add}

\item{reorder}{?? TODO: Zouter/wouters}
}
\description{
Plot strip
}
