% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Visualise a trajectory by the visualisation specific for the method that created it}
\usage{
plot_trajectory(prediction, method, ...)
}
\arguments{
\item{prediction}{The trajectory model to plot}

\item{method}{The method that generated the trajectory.}

\item{...}{Extra parameters for the plotting function.}
}
\description{
Visualise a trajectory by the visualisation specific for the method that created it
}
