% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priors}
\alias{priors}
\title{Metadata on priors}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.}
\usage{
priors
}
\description{
Metadata on priors
}
\keyword{datasets}
