% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dynplot.R
\name{process_dynplot}
\alias{process_dynplot}
\title{Default theme for TI plots}
\usage{
process_dynplot(g, id = NULL, expand = TRUE)
}
\arguments{
\item{g}{A ggplot to modify}

\item{id}{The title}

\item{expand}{Whether or not to leave space at the borders}
}
\description{
Default theme for TI plots
}
