% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{project_waypoints}
\alias{project_waypoints}
\title{Project the waypoints}
\usage{
project_waypoints(traj, cell_positions,
  waypoints = dynwrap::select_waypoints(traj),
  trajectory_projection_sd = sum(traj$milestone_network$length) * 0.05,
  n_arrows = 10)
}
\arguments{
\item{traj}{The trajectory}

\item{cell_positions}{The positions of the cells}

\item{waypoints}{The waypoints to use for projecting, as generated by [dynwrap::select_waypoints()]}

\item{trajectory_projection_sd}{The standard deviation of the gaussian kernel}

\item{n_arrows}{Number of arrows to add to trajectory projection}
}
\description{
Project the waypoints
}
