% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{questions}
\alias{questions}
\title{The different questions}
\format{An object of class \code{list} of length 13.}
\usage{
questions
}
\description{
The different questions
}
\keyword{datasets}
