% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{renderers}
\alias{renderers}
\title{Renderers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 9 columns.}
\usage{
renderers
}
\description{
Renderers
}
\keyword{datasets}
