% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scorpius.R
\name{ti_scorpius}
\alias{ti_scorpius}
\alias{ti_scorpius_sparse}
\title{Inferring trajectories with SCORPIUS}
\usage{
ti_scorpius(distance_method = "spearman", ndim = 3L, k = 4L,
  thresh = 0.001, maxit = 10L, stretch = 0, smoother = "smooth.spline",
  sparse = FALSE)

ti_scorpius_sparse(distance_method = "spearman", ndim = 3L, k = 4L,
  thresh = 0.001, maxit = 10L, stretch = 0, smoother = "smooth.spline",
  sparse = TRUE)
}
\arguments{
\item{distance_method}{A character string indicating which correlation
coefficient (or covariance) is to be computed. One of "pearson", "kendall", or "spearman".}

\item{ndim}{The number of dimensions in the new space.}

\item{k}{The number of clusters to cluster the data into.}

\item{thresh}{\code{\link[princurve]{principal.curve}} parameter: convergence threshhold on shortest distances to the curve}

\item{maxit}{\code{\link[princurve]{principal.curve}} parameter: maximum number of iterations}

\item{stretch}{\code{\link[princurve]{principal.curve}} parameter: a factor by which the curve can be extrapolated when points are projected}

\item{smoother}{\code{\link[princurve]{principal.curve}} parameter: choice of smoother}

\item{sparse}{Whether or not to use sparse MDS dimensionality reduction,
for datasets with large amounts of cells.}
}
\description{
Inferring trajectories with SCORPIUS
}
