% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_waypoints_to_wrapper.R
\name{select_waypoints}
\alias{select_waypoints}
\title{Select the waypoints}
\usage{
select_waypoints(traj, n_waypoints = 100,
  resolution = sum(traj$milestone_network$length)/n_waypoints)
}
\arguments{
\item{traj}{Wrapper with trajectory}

\item{n_waypoints}{The number of waypoints}

\item{resolution}{The resolution of the waypoints, measured in the same units as the lengths of the milestone network edges, will be automatically computed using n_waypoints}
}
\description{
Waypoints are spread equally over the whole trajectory
}
