% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_set_cores.R
\name{set_cores}
\alias{set_cores}
\title{Helper function for nasty Python code}
\usage{
set_cores(n_cores = 1)
}
\arguments{
\item{n_cores}{the number of cores Python is allowed to use.}
}
\description{
Helper function for nasty Python code
}
