% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R
\name{ti_angle}
\alias{ti_angle}
\title{Inferring trajectories with Angle}
\usage{
ti_angle(dimred = "pca")
}
\arguments{
\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}
}
\description{
Inferring trajectories with Angle
}
