% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_comp1.R
\name{ti_comp1}
\alias{ti_comp1}
\title{Inferring trajectories with Component 1}
\usage{
ti_comp1(dimred = "pca", ndim = 2L)
}
\arguments{
\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}

\item{ndim}{The number of dimensions}
}
\description{
Inferring trajectories with Component 1
}
