% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_gpfates.R
\name{ti_gpfates}
\alias{ti_gpfates}
\title{Inferring trajectories with GPfates}
\usage{
ti_gpfates(log_expression_cutoff = 2, min_cells_expression_cutoff = 2,
  ndim = 2L)
}
\arguments{
\item{log_expression_cutoff}{The log expression cutoff}

\item{min_cells_expression_cutoff}{The min expression cutoff}

\item{ndim}{The number of dimensions}
}
\description{
Inferring trajectories with GPfates
}
