% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_identity.R
\name{ti_identity}
\alias{ti_identity}
\title{Inferring trajectories with Control: identity}
\usage{
ti_identity(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{This parameter does not do anything.}
}
\description{
This control method will return the gold standard.
}
