% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_matcher.R
\name{ti_matcher}
\alias{ti_matcher}
\title{Inferring trajectories with MATCHER}
\usage{
ti_matcher(quantiles = 50L, method = "linear")
}
\arguments{
\item{quantiles}{How many quantiles to use when computing warp functions (integer)}

\item{method}{Gaussian process regression or linear interpolation? ("gp" or "linear)}
}
\description{
Inferring trajectories with MATCHER
}
