% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_merlot.R
\name{ti_merlot}
\alias{ti_merlot}
\title{Inferring trajectories with MERLoT}
\usage{
ti_merlot(sigma = "local", distance = "euclidean", ndim = 20L,
  density_norm = TRUE, n_local_lower = 5L, n_local_upper = 7L,
  w_width = 0.1, n_components_to_use = 3L, N_yk = 100L,
  lambda_0 = 4.53999297624849e-05, mu_0 = 0.0025, increaseFactor_mu = 20,
  increaseFactor_lambda = 20, FixEndpoints = FALSE)
}
\arguments{
\item{sigma}{Diffusion scale parameter of the Gaussian kernel. A larger sigma might be necessary if the eigenvalues can not be found because of a singularity in the matrix. Must be one of:
\itemize{
\item{A character vector: \code{"local"} (default) or \code{"global"},}
\item{a numeric global sigma -- a global sigma will be calculated using \code{\link[destiny]{find_sigmas}}}
\item{or a \code{\link[destiny]{Sigmas-class}} object.}
}}

\item{distance}{A \code{\link[stats]{dist}} object, or a character vector specifying which distance metric to use. Allowed measures:
\itemize{
\item{Euclidean distance (default),}
\item{cosine distance (1-corr(c_1, c_2)), or}
\item{rank correlation distance (1-corr(rank(c_1), rank(c_2)))}
}}

\item{ndim}{Number of eigenvectors/dimensions to return}

\item{density_norm}{logical. If TRUE, use density normalisation}

\item{n_local_lower}{If sigma == 'local', the \code{n_local_lower}:\code{n_local_upper} nearest neighbor(s) determine(s) the local sigma}

\item{n_local_upper}{See \code{n_local_lower}}

\item{w_width}{Window width to use for deciding the branch cutoff}

\item{n_components_to_use}{Which components to use in downstream analysis}

\item{N_yk}{number of nodes for the elastic principal tree}

\item{lambda_0}{principal elastic tree energy function parameter.}

\item{mu_0}{principal elastic tree energy function parameter.}

\item{increaseFactor_mu}{factor by which the mu will be increased for the embedding}

\item{increaseFactor_lambda}{factor by which the mu will be increased for the embedding}

\item{FixEndpoints}{Documentation not provided by authors}
}
\description{
Inferring trajectories with MERLoT
}
