% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouija.R
\name{ti_ouija}
\alias{ti_ouija}
\title{Inferring trajectories with ouija}
\usage{
ti_ouija(iter = 100, response_type = "switch", inference_type = "hmc",
  normalise_expression = TRUE)
}
\arguments{
\item{iter}{Number of iterations}

\item{response_type}{A vector declaring whether each gene exhibits "switch" or "transient"
expression. Defaults to "switch" for all genes}

\item{inference_type}{The type of inference to be performed, either \code{hmc} for Hamiltonian
Monte Carlo or \code{vb} for ADVI (Variational Bayes). Note that HMC is typically more accurate
but VB will be orders of magnitude faster.}

\item{normalise_expression}{Logical, default TRUE. If TRUE the data is pre-normalised
so the average peak expression is approximately 1. This makes the strength parameters
approximately comparable between genes.}
}
\description{
Inferring trajectories with ouija
}
\seealso{
\code{\link[ouija:ouija]{ouija::ouija()}}
}
