% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouijaflow.R
\name{ti_ouijaflow}
\alias{ti_ouijaflow}
\title{Inferring trajectories with ouijaflow}
\usage{
ti_ouijaflow(iter = 1000)
}
\arguments{
\item{iter}{Number of iterations}
}
\description{
Inferring trajectories with ouijaflow
}
