% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_periodpc.R
\name{ti_periodpc}
\alias{ti_periodpc}
\title{Inferring trajectories with Periodic PrinCurvee}
\usage{
ti_periodpc(ndim = 3L, maxit = 10L)
}
\arguments{
\item{ndim}{The number of dimensions}

\item{maxit}{maximum number of iterations.}
}
\description{
Inferring trajectories with Periodic PrinCurvee
}
\seealso{
\code{\link[princurve:principal.curve]{princurve::principal.curve()}}
}
