% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pseudogp.R
\name{ti_pseudogp}
\alias{ti_pseudogp}
\title{Inferring trajectories with pseudogp}
\usage{
ti_pseudogp(smoothing_alpha = 10, smoothing_beta = 3,
  pseudotime_mean = 0.5, pseudotime_var = 1, chains = 3L, iter = 100,
  dimreds = c(TRUE, TRUE, FALSE, FALSE, FALSE), initialise_from = "random")
}
\arguments{
\item{smoothing_alpha}{The hyperparameter for the Gamma distribution that controls arc-length}

\item{smoothing_beta}{The hyperparameter for the Gamma distribution that controls arc-length}

\item{pseudotime_mean}{The mean of the constrained normal prior on the pseudotimes}

\item{pseudotime_var}{The variance of the constrained normal prior on the pseudotimes}

\item{chains}{The number of chains for the MCMC trace}

\item{iter}{The number of iterations for the MCMC trace}

\item{dimreds}{A character vector specifying which dimensionality reduction methods to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}

\item{initialise_from}{How to initialise the MCMC chain. One of "random" (stan decides),
"principal_curve", or "pca" (the first component of PCA rescaled is taken to be the pseudotimes).
Note: if multiple representations are provided, \code{pseudogp} will take the principal curve or
pca from the first rather than combining them. If a particular representation is required, it is
up to the user to re-order them.}
}
\description{
Inferring trajectories with pseudogp
}
\seealso{
\code{\link[pseudogp:fitPseudotime]{pseudogp::fitPseudotime()}}
}
