% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_random.R
\name{ti_random}
\alias{ti_random}
\title{Inferring trajectories with Control: random}
\usage{
ti_random(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{This parameter does not do anything.}
}
\description{
This control method will always produce a random trajectory.
}
