% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scuba.R
\name{ti_scuba}
\alias{ti_scuba}
\title{Inferring trajectories with SCUBA}
\usage{
ti_scuba(rigorous_gap_stats = TRUE, N_dim = 2L, low_gene_threshold = 1,
  low_gene_fraction_max = 0.7, min_split = 15L,
  min_percentage_split = 0.25)
}
\arguments{
\item{rigorous_gap_stats}{Whether to use rigorous gap statistics to determine number of clusters}

\item{N_dim}{Number of TSNE dimensions}

\item{low_gene_threshold}{Threshold value for genes of low expression levels}

\item{low_gene_fraction_max}{Maximum fraction of lowly-expressed cells allowed for each gene}

\item{min_split}{Lower threshold on the number of cells in a cluster for this cluster to be split.}

\item{min_percentage_split}{Minimum fraction of cells in the smaller cluster during a bifurcation.}
}
\description{
Inferring trajectories with SCUBA
}
