% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_shuffle.R
\name{ti_shuffle}
\alias{ti_shuffle}
\title{Inferring trajectories with Control: shuffle}
\usage{
ti_shuffle(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{This parameter does not do anything.}
}
\description{
This control method will return the milestone network of the provided
gold standard, but will shuffle the cell positions randomly.
}
