% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slicer.R
\name{ti_slicer}
\alias{ti_slicer}
\title{Inferring trajectories with SLICER}
\usage{
ti_slicer(kmin = 10L, m = 2L)
}
\arguments{
\item{kmin}{Smallest value of k to try}

\item{m}{ intrinsic dimension of the data. This parameter mainly influences the visualisation of the results. The real intrinsic dimension will be calculated automaticly. }
}
\description{
Inferring trajectories with SLICER
}
