% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slingshot.R
\name{ti_slingshot}
\alias{ti_slingshot}
\title{Inferring trajectories with Slingshot}
\usage{
ti_slingshot(ndim = 3L, nclus = 5L, dimred = "pca", shrink = 1,
  reweight = TRUE, drop.multi = TRUE, thresh = 0.001, maxit = 10L,
  stretch = 2, smoother = "smooth.spline", shrink.method = "cosine")
}
\arguments{
\item{ndim}{The number of dimensions}

\item{nclus}{Number of clusters}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}

\item{shrink}{logical or numeric between 0 and 1, determines whether and how 
much to shrink branching lineages toward their average prior to the split.}

\item{reweight}{logical, whether to allow cells shared between lineages to be
reweighted during curve-fitting. If \code{TRUE}, cells shared between 
lineages will be weighted by: distance to nearest curve / distance to
curve.}

\item{drop.multi}{logical, whether to drop shared cells from lineages which
do not fit them well. If \code{TRUE}, shared cells with a distance to one 
lineage above the 90th percentile and another below the 50th will be
dropped from the further lineage.}

\item{thresh}{numeric, determines the convergence criterion. Percent change
in the total distance from cells to their projections along curves must be
less than \code{thresh}. Default is \code{0.001}, similar to 
\code{\link[princurve]{principal.curve}}.}

\item{maxit}{numeric, maximum number of iterations, see 
\code{\link[princurve]{principal.curve}}.}

\item{stretch}{numeric factor by which curves can be extrapolated beyond 
endpoints. Default is \code{2}, see \code{\link[princurve]{principal.curve}}.}

\item{smoother}{choice of scatter plot smoother. Same as 
\code{\link[princurve]{principal.curve}}, but \code{"lowess"} option is replaced with 
\code{"loess"} for additional flexibility.}

\item{shrink.method}{character denoting how to determine the appropriate
amount of shrinkage for a branching lineage. Accepted values are the same
as for \code{kernel} in \code{\link{density}} (default is \code{"cosine"}),
as well as \code{"tricube"} and \code{"density"}. See 'Details' for more.}
}
\description{
Inferring trajectories with Slingshot
}
\seealso{
\code{\link[slingshot:slingshot]{slingshot::slingshot()}}
}
