% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_topslam.R
\name{ti_topslam}
\alias{ti_topslam}
\title{Inferring trajectories with topslam}
\usage{
ti_topslam(n_components = 2L, n_neighbors = 10L, linear_dims = 0L,
  max_iters = 1000, dimreds = c(TRUE, TRUE, TRUE, TRUE, TRUE))
}
\arguments{
\item{n_components}{the number of components}

\item{n_neighbors}{the number of neighbors}

\item{linear_dims}{?}

\item{max_iters}{the number of iterations to optimize over}

\item{dimreds}{A character vector specifying which dimensionality reduction methods to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}
}
\description{
Inferring trajectories with topslam
}
