% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_tscan.R
\name{ti_tscan}
\alias{ti_tscan}
\title{Inferring trajectories with TSCAN}
\usage{
ti_tscan(minexpr_percent = 0, minexpr_value = 0, cvcutoff = 0,
  clusternum_lower = 2L, clusternum_upper = 9L, modelNames = "VVV")
}
\arguments{
\item{minexpr_percent}{Numeric value specifying the lowest percentage of highly expressed cells (expression value bigger than minexpr_value) for the genes/features to be retained.}

\item{minexpr_value}{Numeric value specifying the minimum cutoff of log transformed (if takelog is TRUE) value}

\item{cvcutoff}{Numeric value specifying the minimum value of coefficient of variance for the genes/features to be retained.}

\item{clusternum_lower}{An integer specifying the minimal possible cluster numbers. The best cluster number will be picked using BIC.}

\item{clusternum_upper}{An integer specifying the maximal possible cluster numbers. The best cluster number will be picked using BIC.}

\item{modelNames}{model to be used in model-based clustering. By default "ellipsoidal, varying volume, shape, and orientation" is used.}
}
\description{
Inferring trajectories with TSCAN
}
