% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_waterfall.R
\name{ti_waterfll}
\alias{ti_waterfll}
\title{Inferring trajectories with Waterfall}
\usage{
ti_waterfll(num_clusters = 10L)
}
\arguments{
\item{num_clusters}{Number of clusters}
}
\description{
Inferring trajectories with Waterfall
}
