% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_data.R
\name{wrap_data}
\alias{wrap_data}
\title{A data wrapper for datasets and trajectories}
\usage{
wrap_data(id, cell_ids, cell_info = NULL, ...)
}
\arguments{
\item{id}{A unique identifier for the data}

\item{cell_ids}{The ids of the cells.}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{...}{Extra information to be stored in the wrapper.}
}
\description{
A data wrapper for datasets and trajectories
}
