% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_feature_importance.R
\name{calculate_milestone_feature_importance}
\alias{calculate_milestone_feature_importance}
\alias{calculate_overall_feature_importance}
\alias{calculate_waypoint_feature_importance}
\alias{calculate_cell_feature_importance}
\alias{calculate_branch_feature_importance}
\alias{calculate_branching_point_feature_importance}
\title{Calculating feature importances across trajectories}
\usage{
calculate_milestone_feature_importance(traj, expression_source = "expression",
  milestones_oi = NULL, method = "ranger", method_params = list())

calculate_overall_feature_importance(traj, expression_source = "expression",
  method = "ranger", method_params = list())

calculate_waypoint_feature_importance(traj, expression_source = "expression",
  waypoints = NULL, method = "ranger", method_params = list())

calculate_cell_feature_importance(traj, expression_source = "expression",
  method = "ranger", method_params = list())

calculate_branch_feature_importance(traj, expression_source = "expression",
  method = "ranger", method_params = list())

calculate_branching_point_feature_importance(traj,
  expression_source = "expression", milestones_oi = traj$milestone_ids,
  method = "ranger", method_params = list())
}
\arguments{
\item{traj}{A trajectory object containing expression values and a trajectory.}

\item{expression_source}{The expression_source, if not provided will use the expression within the trajectory}

\item{milestones_oi}{The milestone(s) for which to calculate feature importance}

\item{method}{The method to do regressions, can be `ranger` or any regression model from caret}

\item{method_params}{Parameters given to the method}

\item{waypoints}{The waypoints, optional}
}
\description{
Uses the feature importance measures of \code{\link[ranger]{ranger}} or  \code{caret}. \code{calculate_overall_feature_importance} calculates the importance for the whole trajectory, \code{calculate_milestone_feature_importance} calculates it for individual milestones (eg. branching points)
}
