% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_dimred.R
\name{dimred_trajectory}
\alias{dimred_trajectory}
\title{Perform dimensionality reduction on a trajectory and the respective samples in order to plot it}
\usage{
dimred_trajectory(traj)
}
\arguments{
\item{traj}{A trajectory as generated by
\code{\link[dynwrap]{wrap_data}} and \code{\link[dynwrap]{add_trajectory_to_wrapper}}.}
}
\description{
Perform dimensionality reduction on a trajectory and the respective samples in order to plot it
}
