% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_infer_trajectory.R
\name{execute_method_internal}
\alias{execute_method_internal}
\title{Internal method for executing a method}
\usage{
execute_method_internal(method, arglist, setseed_detection_file)
}
\arguments{
\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (e.g. \code{\link{ti_paga}()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}), or}
\item{a dynguidelines data frame.}
}}

\item{arglist}{The arguments to apply to the method}

\item{setseed_detection_file}{A file to which will be written if a method
uses the set.seed function.}
}
\description{
If you're reading this, you're supposed to be using \code{infer_trajectory} instead.
}
