% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_infer_trajectory.R
\name{execute_method_on_task}
\alias{execute_method_on_task}
\title{Run a method on a task with a set of parameters}
\usage{
execute_method_on_task(task, method, parameters = list(),
  give_priors = NULL, mc_cores = 1, verbose = FALSE)
}
\arguments{
\item{task}{The task}

\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (e.g. \code{\link{ti_paga}()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}), or}
\item{a dynguidelines data frame.}
}}

\item{parameters}{A set of parameters to be used during trajectory inference.
A parameter set must be a named list of parameters.
If multiple methods were provided in the \code{method} parameter,
\code{parameters} must be an unnamed list of the same length.}

\item{give_priors}{All the priors a method is allowed to receive. Must be a subset of: \code{"start_milestones"},
\code{"start_cells"}, \code{"end_milestones"}, \code{"end_cells"}, \code{"grouping_assignment"} and \code{"grouping_network"}}

\item{mc_cores}{The number of cores to use, allowing to parallellise the different tasks}

\item{verbose}{Whether or not to print information output}
}
\description{
Run a method on a task with a set of parameters
}
