% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{dimred_pca}
\alias{dimred_pca}
\alias{dimred_mds}
\alias{dimred_mds_sammon}
\alias{dimred_mds_isomds}
\alias{dimred_mds_smacof}
\alias{dimred_tsne}
\alias{dimred_dp}
\alias{dimred_ica}
\alias{dimred_lle}
\alias{dimred_umap}
\alias{get_dimreds}
\title{Dimensionality reduction functions}
\usage{
dimred_pca(x, ndim = 3)

dimred_mds(x, ndim = 3)

dimred_mds_sammon(x, ndim = 3)

dimred_mds_isomds(x, ndim = 3)

dimred_mds_smacof(x, ndim = 3)

dimred_tsne(x, ndim = 3)

dimred_dp(x, ndim = 3)

dimred_ica(x, ndim = 3)

dimred_lle(x, ndim = 3)

dimred_umap(x, ndim = 2, n_neighbors = 15L)

get_dimreds()
}
\arguments{
\item{x}{Data matrix}

\item{ndim}{Number of dimensions}

\item{n_neighbors}{numeric. The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100.}
}
\description{
Dimensionality reduction functions
}
