% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot trajectory on dimensionality reduction}
\usage{
plot_dimred(traj, color_cells = c("auto", "invisible", "grouping", "feature",
  "milestone", "pseudotime"), grouping_assignment = NULL, groups = NULL,
  feature_oi = NULL, color_milestones = c("auto", "given"),
  milestones = NULL, milestone_percentages = NULL, pseudotime = NULL,
  expression_source = "expression",
  plot_milestone_network = dynwrap::is_wrapper_with_trajectory(traj),
  plot_milestone_labels = FALSE,
  dimred_method = ifelse(length(traj$cell_ids) > 500, dimred_pca, dimred_mds))
}
\arguments{
\item{traj}{The traj}

\item{color_cells}{How to color the cells}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the expression}

\item{plot_milestone_network}{Whether to plot the milestone network}

\item{plot_milestone_labels}{Whether to overlay the milestone labels over the network}

\item{dimred_method}{The dimred method or the dimensionality reduction (a dataframe containing at least Comp1, Comp2 and cell_id)}
}
\description{
Plot trajectory on dimensionality reduction
}
