% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_features}
\alias{plot_features}
\title{Plotting the features}
\usage{
plot_features(traj, features_oi = sample(colnames(traj$counts), min(c(5,
  ncol(traj$counts)))), margin = 0.05)
}
\arguments{
\item{traj}{The traj}

\item{features_oi}{The features of interest}

\item{margin}{The margin to add}
}
\description{
Plotting the features
}
