% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the traj as a heatmap}
\usage{
plot_heatmap(traj, expression_source = "expression", features_oi = 20,
  clust = "ward.D2", margin = 0.02, color_cells = NULL,
  milestones = NULL, milestone_percentages = traj$milestone_percentages,
  grouping_assignment = NULL, groups = NULL,
  cell_feature_importances = NULL, heatmap_type = c("tiled", "dotted"))
}
\arguments{
\item{traj}{The traj}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{features_oi}{features to plot, or the top number of features to select}

\item{clust}{The method to cluster the features, or a hclust object}

\item{color_cells}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{cell_feature_importances}{The feature importances per cell}

\item{heatmap_type}{The type of heatmap, either tiled or dotted}
}
\description{
Plot the traj as a heatmap
}
