% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linearised_comparison.R
\name{plot_linearised_comparison}
\alias{plot_linearised_comparison}
\title{Plot strip onedim}
\usage{
plot_linearised_comparison(traj1, traj2, reorder = TRUE, margin = 0.05,
  reorder_second_by = "mapping")
}
\arguments{
\item{traj1}{The first traj}

\item{traj2}{The second traj}

\item{reorder}{?? TODO: Zouter/wouters}

\item{margin}{The margin to add}

\item{reorder_second_by}{?? TODO: Zouter/wouters}
}
\description{
Plot strip onedim
}
