% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_onedim.R
\name{plot_onedim}
\alias{plot_onedim}
\title{Plot onedim}
\usage{
plot_onedim(traj, color_cells = c("auto", "none", "grouping", "feature",
  "milestone", "pseudotime"), milestone_network = traj$milestone_network,
  progressions = traj$progressions, grouping_assignment = NULL,
  groups = NULL, feature_oi = NULL, pseudotime = NULL,
  expression_source = "expression", color_milestones = c("auto", "given"),
  milestones = NULL, milestone_percentages = NULL, orientation = 1,
  margin = 0.05, linearised = linearise_cells(milestone_network,
  progressions, margin, one_edge = TRUE), quasirandom_width = 0.2,
  plot_cells = TRUE,
  label_milestones = dynwrap::is_wrapper_with_milestone_labelling(traj))
}
\arguments{
\item{traj}{The trajectory}

\item{color_cells}{How to color the cells}

\item{milestone_network}{Optional, the milestone network}

\item{progressions}{The progressions used to put the cells on the graph}

\item{grouping_assignment}{Tibble containing the assignment of cells to groups of cells}

\item{groups}{Tibble containing information of the cell groups}

\item{feature_oi}{feature to plot expression}

\item{pseudotime}{The pseudotime}

\item{expression_source}{Source of the feature expression, defaults to `expression`}

\item{color_milestones}{How to color the cells}

\item{milestones}{Tibble containing the `milestone_id` and a `color` for each milestone}

\item{milestone_percentages}{The milestone percentages}

\item{orientation}{Whether to plot the connections in the top (1) or bottom (-1)}

\item{margin}{The margin to add}

\item{linearised}{The linearised milestone network and progressions}

\item{quasirandom_width}{The width of the quasirandom cell spreading}

\item{plot_cells}{Whether to plot the cells}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels as milestone_ids will be used), a named character vector, or FALSE}
}
\description{
Plot onedim
}
