% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_docker.R
\name{start_dynmethods_docker}
\alias{start_dynmethods_docker}
\title{Start the dynmethods docker}
\usage{
start_dynmethods_docker()
}
\description{
Make sure docker is installed and available through path
}
